pyproject.toml
[project]
name = "highload_2025_1"
version = "0.1.0"
description = "test project"
readme = "readme.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "PyJWT",
    "pydantic",
    "pytest",
    "aiopg",
]
